+++ packages/server/src/command-server.ts
@@ -1,15 +1,14 @@
-import { Operation, spawn } from 'effection';
+import { Operation, spawn, fork } from 'effection';
 import { Mailbox } from '@bigtest/effection';
 import { express } from '@bigtest/effection-express';
 import * as graphqlHTTP from 'express-graphql';
 import { graphql as executeGraphql } from 'graphql';
 
-import { listenWS } from './ws';
+import { Connection, sendData, listenWS } from './ws';
 import { schema } from './schema';
 import { Atom } from '@bigtest/atom';
 import { OrchestratorState } from './orchestrator/state';
-
-import { handleMessage } from './command-server/websocket';
+import { Message, QueryMessage, MutationMessage, isQuery, isMutation } from './protocol';
 
 export type CommandMessage = { status: "ready" } | { type: "run"; id: string };
 
@@ -39,7 +38,7 @@ export function* createCommandServer(options: CommandServerOptions): Operation {
  * Run the query or mutation in `source` against the orchestrator
  * state contained in `state`
  */
-export function graphql(source: string, delegate: Mailbox, state: OrchestratorState): Operation {
+function graphql(source: string, delegate: Mailbox, state: OrchestratorState): Operation {
   let options = graphqlOptions(delegate, state);
   return executeGraphql({...options, contextValue: options.context, source });
 }
@@ -55,10 +54,52 @@ let testRunIds = (function * () {
  * because the express graphql server calls the `graphql` function for
  * you based on the .
  */
-export function graphqlOptions(delegate: Mailbox, state: OrchestratorState) {
+function graphqlOptions(delegate: Mailbox, state: OrchestratorState) {
   return {
     schema,
     rootValue: state,
     context: { delegate, testRunIds }
   };
+}
+
+function handleMessage(delegate: Mailbox, atom: Atom<OrchestratorState>): (connection: Connection) => Operation {
+  function* handleQuery(message: QueryMessage, connection: Connection): Operation {
+    yield publishQueryResult(message, atom.get(), connection);
+
+    if (message.live) {
+      yield fork(subscribe(message, connection));
+    }
+  }
+
+  function* handleMutation(message: MutationMessage, connection: Connection): Operation {
+    let result = yield graphql(message.mutation, delegate, atom.get());
+    result.responseId = message.responseId;
+    yield sendData(connection, JSON.stringify(result));
+  }
+
+  function* publishQueryResult(message: QueryMessage, state: OrchestratorState, connection: Connection): Operation {
+    let result = yield graphql(message.query, delegate, state);
+    result.responseId = message.responseId;
+    yield sendData(connection, JSON.stringify(result));
+  }
+
+  function* subscribe(message: QueryMessage, connection: Connection) {
+    yield atom.each(state => publishQueryResult(message, state, connection));
+  }
+
+  return function*(connection) {
+    let messages: Mailbox = yield Mailbox.subscribe(connection, "message");
+
+    while (true) {
+      let { args } = yield messages.receive();
+      let message = JSON.parse(args[0].utf8Data) as Message;
+
+      if (isQuery(message)) {
+        yield fork(handleQuery(message, connection));
+      }
+      if (isMutation(message)) {
+        yield fork(handleMutation(message, connection));
+      }
+    }
+  }
 }
